requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0  # для работы с изображениями
python-dotenv>=1.0.0  # для работы с переменными окружения
pytest>=7.4.0  # для тестирования
pytest-django>=4.5.2  # для тестирования Django
psycopg2-binary>=2.9.9  # для работы с PostgreSQL (опционально) 